Fix: Decimales precio Ejemplar
Corregido el tipo de dato del precio de los ejemplares
Las paginas de creación y actualización de
ejemplares ahora soportan precios decimales

+++ scripts/1 - almacen_db.sql
@@ -51,7 +51,7 @@ CREATE TABLE IF NOT EXISTS Ejemplares (
     producto_fk INT NOT NULL,
     ubicacion_fk INT NOT NULL,
     venta_fk INT,
-    precio INT NOT NULL,
+    precio DECIMAL(14, 2) NOT NULL,
     fecha_entrada DATE,
     fecha_actualizacion DATE,
     concurrencia INT,
